backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary>=2.9
django-cors-headers>=3.13
channels>=4.0
channels-redis>=4.3
celery[redis]>=5.3
django-filter>=23.1
drf-spectacular>=0.28
python-dotenv>=1.0
gunicorn>=20.1
pytest-django>=4.5
daphne
